+++ directional_pool.go
@@ -20,8 +20,8 @@ type DirectionalPool struct {
 	ResultURI                 string `json:"resultUri"`
 }
 
-// AccountLevelGeoDirectionalGroupDTO wraps an account-level, geo directonal-group response
-type AccountLevelGeoDirectionalGroupDTO struct {
+// AccountLevelGeoDirectionalGroup wraps an account-level, geo directonal-group response
+type AccountLevelGeoDirectionalGroup struct {
 	Name        string   `json:"name"`
 	Description string   `json:"description"`
 	Codes       []string `json:"codes"`
@@ -51,10 +51,10 @@ type DirectionalPoolListDTO struct {
 
 // AccountLevelGeoDirectionalGroupListDTO wraps a list of account-level, geo directional-groups response from a index request
 type AccountLevelGeoDirectionalGroupListDTO struct {
-	AccountName string                               `json:"zoneName"`
-	GeoGroups   []AccountLevelGeoDirectionalGroupDTO `json:"geoGroups"`
-	Queryinfo   QueryInfo                            `json:"queryInfo"`
-	Resultinfo  ResultInfo                           `json:"resultInfo"`
+	AccountName string                            `json:"zoneName"`
+	GeoGroups   []AccountLevelGeoDirectionalGroup `json:"geoGroups"`
+	Queryinfo   QueryInfo                         `json:"queryInfo"`
+	Resultinfo  ResultInfo                        `json:"resultInfo"`
 }
 
 // AccountLevelIPDirectionalGroupListDTO wraps an account-level, IP directional-group response
@@ -134,13 +134,13 @@ type GeoDirectionalPoolsService struct {
 }
 
 // Select requests all geo directional-pools, by query and account, providing pagination and error handling
-func (s *GeoDirectionalPoolsService) Select(k GeoDirectionalPoolKey, query string) ([]AccountLevelGeoDirectionalGroupDTO, error) {
+func (s *GeoDirectionalPoolsService) Select(k GeoDirectionalPoolKey, query string) ([]AccountLevelGeoDirectionalGroup, error) {
 	// TODO: Sane Configuration for timeouts / retries
 	maxerrs := 5
 	waittime := 5 * time.Second
 
 	// init accumulators
-	dtos := []AccountLevelGeoDirectionalGroupDTO{}
+	dtos := []AccountLevelGeoDirectionalGroup{}
 	errcnt := 0
 	offset := 0
 
@@ -170,12 +170,12 @@ func (s *GeoDirectionalPoolsService) Select(k GeoDirectionalPoolKey, query strin
 }
 
 // SelectWithOffset requests list of geo directional-pools, by query & account, and an offset, returning the directional-group, the list-metadata, the actual response, or an error
-func (s *GeoDirectionalPoolsService) SelectWithOffset(k GeoDirectionalPoolKey, query string, offset int) ([]AccountLevelGeoDirectionalGroupDTO, ResultInfo, *http.Response, error) {
+func (s *GeoDirectionalPoolsService) SelectWithOffset(k GeoDirectionalPoolKey, query string, offset int) ([]AccountLevelGeoDirectionalGroup, ResultInfo, *http.Response, error) {
 	var tld AccountLevelGeoDirectionalGroupListDTO
 
 	res, err := s.client.get(k.QueryURI(query, offset), &tld)
 
-	pis := []AccountLevelGeoDirectionalGroupDTO{}
+	pis := []AccountLevelGeoDirectionalGroup{}
 	for _, pi := range tld.GeoGroups {
 		pis = append(pis, pi)
 	}
@@ -183,8 +183,8 @@ func (s *GeoDirectionalPoolsService) SelectWithOffset(k GeoDirectionalPoolKey, q
 }
 
 // Find requests a geo directional-pool by name & account
-func (s *GeoDirectionalPoolsService) Find(k GeoDirectionalPoolKey) (AccountLevelGeoDirectionalGroupDTO, *http.Response, error) {
-	var t AccountLevelGeoDirectionalGroupDTO
+func (s *GeoDirectionalPoolsService) Find(k GeoDirectionalPoolKey) (AccountLevelGeoDirectionalGroup, *http.Response, error) {
+	var t AccountLevelGeoDirectionalGroup
 	res, err := s.client.get(k.URI(), &t)
 	return t, res, err
 }