+++ udnssdk_test.go
@@ -27,7 +27,7 @@ var (
 	testIPDPoolDescr   = "A Test IP Directional Pool Group"
 	testIPAddrRange    = IPAddrRange{Address: "127.0.0.1"}
 	testIPDPool        = AccountLevelIPDirectionalGroupDTO{Name: "testippool", Description: "An IP Test Pool", IPs: []IPAddrRange{IPAddrRange{Address: "127.0.0.1"}}}
-	testGeoDPool       = AccountLevelGeoDirectionalGroupDTO{Name: "testgeopool", Description: "A test geo pool", Codes: []string{"US, UK"}}
+	testGeoDPool       = AccountLevelGeoDirectionalGroup{Name: "testgeopool", Description: "A test geo pool", Codes: []string{"US, UK"}}
 	testGeoDPoolName   = "testgeodpool"
 	testGeoDPoolDescr  = "A Test Geo Directional Pool Group"
 	testGeoDPoolCodes  = []string{"US", "UK"}